requirements.txt
streamlit==1.28.1
pandas==2.0.3
numpy==1.24.3
plotly==5.17.0
yfinance==0.2.32
requests==2.32.3
google-genai==1.30.0
scipy==1.11.2
scikit-learn==1.3.0
